WIP: Remove warnings
The SPL files now compile without warnings. Biggest change is dropping
the illumos "p0" for epoch-process. It has never been used for anything
but as a unique address identifying the first spa created.

Warning 4146 is currently left disabled due to the P2ALIGN macros.

+++ ZFSin/spl/include/sys/taskq_impl.h
@@ -151,7 +151,7 @@ struct taskq {
 	hrtime_t	tq_totaltime;	/* Time spent processing tasks */
 	uint64_t	tq_tasks;	/* Total # of tasks posted */
 	uint64_t	tq_executed;	/* Total # of tasks executed */
-	int		tq_maxtasks;	/* Max number of tasks in the queue */
+	uint64_t		tq_maxtasks;	/* Max number of tasks in the queue */
 	int		tq_tcreates;
 	int		tq_tdeaths;
 };
